Changed Add signature

+++ GitTest/Calculator.cs
@@ -8,7 +8,7 @@ namespace GitTest
 {
     public static class Calculator
     {
-        public static int Add(int x, int y)
+        public static double Add(double x, double y)
         {
             return x + y;
         }
